cleaning & merging
