+++ timpute/test/test_do.py
@@ -4,8 +4,6 @@ from ..impute_helper import calcR2X
 from ..generateTensor import generateTensor
 from timpute.tracker import Tracker
 from timpute.decomposition import Decomposition
-from timpute.common import getSetup
-from timpute.plot import q2xentry
 
 def test_decomp_dopt(plot=False, method = perform_DO):
     tensor = generateTensor('known', shape = (10,10,10))
@@ -17,11 +15,6 @@ def test_decomp_dopt(plot=False, method = perform_DO):
     print("average Q2X: " + np.array2string(np.average(decomp.entry_total, axis=0)))
     print("average fitted Q2X: " + np.array2string(np.average(decomp.entry_fitted, axis=0)))
     print("average imputed Q2X: " + np.array2string(np.average(decomp.entry_imputed, axis=0)))
-
-    if plot:
-        ax, f = getSetup((5,3),(1,1))
-        q2xentry(ax, decomp, method.__name__, detailed=False)
-        return f
     
     return decomp, track
